Actualizacion entidades

+++ src/main/java/com/chocolate/amaro/model/request/ProductRequest.java
@@ -21,12 +21,10 @@ public class ProductRequest {
     @NotBlank(message = "description may not be empty")
     private String description;
 
-   // @NotBlank(message = "image may not be empty")
-    //@Pattern(regexp = "((http|https)://)?(www.)?[a-zA-Z0-9@:%._\\+~#?&//=]{2,256}\\.[a-z]{2,6}\\b([-a-zA-Z0-9@:%._\\+~#?&//=]*)(.jpg|.png|.jpeg)", message="The image URL has invalid format")
     private String image;
 
     @NotNull(message = "Price may not be empty")
-    private Long price;
+    private Double price;
 
     @NotNull(message = "Category may not be empty")
     private long categoryId;
